Update stop sequence to new task format
I missed this in the prior commits, but this updates the stop sequence
passed to the Completions API to match our new task format so that the
model only generates two new instructions per run. We ask it to
generate five, but seed it with two examples, it generates two, and it
stops as soon as it starts generating the fifth.

Signed-off-by: Ben Browning <bbrownin@redhat.com>