Rename cli package and ilab command to instructlab
The `cli` package and PyPI project have been renamed to `instructlab`.
The package is now located in a `src` folder so we can properly test the package.

The CLI is now available as `ilab` and as `instructlab` scripts.

Unit and integration tests are now installing the `instructlab`, so we
are testing what we ship to our users.

Documentation, tests, PlantUML, and comments have been updated to
`instructlab`, too.

Signed-off-by: Christian Heimes <cheimes@redhat.com>