Enforce maximum length in generate seed examples
This calculates a maximum number of characters per seed instruction
example based on the configured server context size as passed to `ilab
generate`. Before starting the generation, we verify that none of the
examples we're going to use exceed this limit. If they do, we fail
generation with a helpful error message.

When adding generated responses to our list of seed examples, we do a
similar check of the response length. If the response is too long, we
skip adding that particular response to our list of seed examples so
that we don't end up adding really long machine-generated responses to
our seed examples and blowing our context size that way.

Fixes #1024

Signed-off-by: Ben Browning <bbrownin@redhat.com>