Makefile: Add target to check for missing updates
This code was pulled in from the `instructlab/instructlab` repo. The
Makefile includes the commit ref that was imported. Running `make
check` will quickly report whether changes have been made to that code
since it was imported.

This is to help make sure we don't miss anything while working through
this code move.

Signed-off-by: Russell Bryant <rbryant@redhat.com>