requirements.txt: instructlab to requirements-dev.txt
This is still a dependency, but it's causing some problems in
`instructlab` repo CI. A released version of instructlab gets
installed first because of this entry, then it has to be uninstalled
before the development version can be installed.

The currently released version of `instructlab` still has schema
embedded. The next release (and current `main`) uses the
`instructlab-schema` package. Moving from one to the other in tox
is breaking the schema package.

By moving it, it will still be installed for local repo testing
purposes.

This should solve the CI failures seen here:
https://github.com/instructlab/instructlab/pull/1244

Signed-off-by: Russell Bryant <rbryant@redhat.com>