Fix formatting to pass the `ruff` checks
Signed-off-by: Russell Bryant <rbryant@redhat.com>

Move test_knowledge.py out of unit tests dir

Move this script so it's not run in unit test automation.

Signed-off-by: Russell Bryant <rbryant@redhat.com>

Ensure configs get included in Python package

We need to have `__init__.py` in this directory for it to get picked
up and included in the Python package so the files are present after
it gets installed.

Signed-off-by: Russell Bryant <rbryant@redhat.com>

Add SPDX-License-Identifier headers to new source files

Signed-off-by: Russell Bryant <rbryant@redhat.com>