Add API to enable disabling batching
Add a `batched` constructor param.

llama-cpp is still the CLI default and doesn't support batching. We
don't know in this code what backend is being used, so for now just
turn off batching. We need to come back around to disabling it only
when we know the default llama-cpp backend is in use.

Signed-off-by: Russell Bryant <rbryant@redhat.com>
Signed-off-by: Mark McLoughlin <markmc@redhat.com>