Fix prompt file paths for an installed library (#67)
* default_flows: Move sdg_base into base Flow class

Each subclass was calculating the root directory of the python package
for finding prompt templates that are embedded in the package. Move
this to the base `Flow` class so we're only doing it in one place.

Signed-off-by: Russell Bryant <rbryant@redhat.com>

* Fix remaining prompt file paths

Special handling is required to find these prompt files from an
installed version of the instructlab-sdg package. This updates the
rest of this file to make use of `self.sdg_base` from the base `Flow`
class which will work from an installed library.

Signed-off-by: Russell Bryant <rbryant@redhat.com>

---------

Signed-off-by: Russell Bryant <rbryant@redhat.com>