filterblock: Don't break when filtering on unexpected data
The previous code in this block did filtering assuming that all
samples had a value that was correct for the type. For example, when
filtering on an integer value, it assumed every row had a valid
integer, where it may instead have garbage.

This change introduces a new helper, _convert_dtype(), which properly
handles this condition. When the conversion fails on a `ValueError`
exception, it treats it as `None` instead of allowing the exception to
be raised up to the caller.

The fix was authored by Shiv in PR #72. I only pulled it out into a
standalone commit.

Signed-off-by: Russell Bryant <rbryant@redhat.com>
Co-authored-by: shiv <shivchander.s30@gmail.com>