Fix reuse of pipelines including IterBlock
When running `sdg.generate()` more than once, a pipeline including an
`IterBlock` would fail because the `block_kwargs` were removed the
first time the block was used. This change removes the special
handling for `IterBlock` because it does not appear to be necessary.

The new code ends up calling the block's constructor with all of the
same arguments, but now `block_kwargs` doesn't disappear.

Signed-off-by: Russell Bryant <rbryant@redhat.com>