generate_data: fix support for multiple leaf nodes
When running generate_data() with a taxonomy with more than one leaf
node, the code previously appended the resulting datasets as if they
behaved like a Python list. They are a `datasets.Dataset` and do not
support the `+` operator.

Instead, keep a list of these datasets. Also update the code that
writes these results to a file to handle the extra level of iteration
now required.

Signed-off-by: Russell Bryant <rbryant@redhat.com>