Move FilterByValue multiprocessing config to PipelineContext
In every use of FilterByValue in the default flows, we use batch_kwargs
to set num_proc=8.

This doesn't appear to be a pipeline author concern, but rather a
runtime parameter which should in future be based on the number of
available CPUs and (perhaps) user configuration.

For now, just move it from batch_kwargs to PipelineContext.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>