llmblock: Set a more reasonable default for num_tokens
This parameter is part of the OpenAI API for creating a chat
completion.

https://platform.openai.com/docs/api-reference/chat/create#chat-create-max_tokens

In our default flows, we configure many LLMBlocks explicitly to use
`max_tokens` of 2048, but many do not have it set and rely on the
default here.

In some testing with vllm, it was observed that this caused a
significant performance hit to vllm's cache. A value this high is
unnecessary, as we have observed that most of our responses are closer
to 2k, anyway.

A follow-up to this would be to look closer at which blocks we are
explicitly setting to use 2k, and which we are allowing to use the
larger default. It's not clear to me if they are all intentional or
not. Tracked via #126.

This change originally comes from https://github.com/aakankshaduggal/sdg/pull/3/commits/04e33982dd8fba05f52700a9eedccdb115609e80

Co-authored-by: shiv <shivchander.s30@gmail.com>
Co-authored-by: Aakanksha Duggal <aduggal@redhat.com>
Co-authored-by: Kai Xu <xuk@ibm.com>
Signed-off-by: Russell Bryant <rbryant@redhat.com>