Add jsonschema for validating pipeline configuration
This commit introduces a `v1.json` jsonschema for use with validating
pipeline configuration. This follows the same format used in the
`instructlab/schema` repository which we use for validating taxonomy
`qna.yaml` files.

It's very likely that the schema is missing some details that are
allowed. It does pass against all configurations currently in the
source tree, though.

Signed-off-by: Russell Bryant <rbryant@redhat.com>