Update grounded_skills.yaml to add seed value
This commit adds seed parameter to gen_context block in the grounded skills flow.

What is a seed and why we need it?
Seed: This parameter is used to set a seed value for the random number generator used by the model. This allows you to generate reproducible results by providing the same seed value each time.

If seed is not provided, or the model changes, or the input changes (by as much as a single token) then the result will be a completely new result. It's only there for reproducibility.

Signed-off-by: Aakanksha Duggal <aduggal@redhat.com>