filterblock: add default_value for use  with convert_dtype
Fixes #133

We have been using None as a default value when dtype conversion fails.
This works fine with the eq operator, but not others e.g. ge

In many cases a default like zero or False works fine, but there are
cases where the pipeline author will need to specify a different default,
for example when using something like `le(1)` you would want a default
greater than 1.

With this change, we're acknowledging that this is default value
handling is very normal, and we can drop from error logging to debug
logging.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>