schema: Allow `seed` in `gen_kwargs`
A previous commit added `seed` here, which I expected to fail schema
validation. It didn't fail because we did not specify that additional
unspecified parameters are not allowed in this section.

This commit adds `seed` to the schema and adjusts the setting for this
section so that other parameters not specified here will cause
validation to fail.

Signed-off-by: Russell Bryant <rbryant@redhat.com>