docs: Add a doc to track pipeline config version history
Our intiial pipeline configuration is version `1.0`. We haven't
actually released anything yet, but I think it would be good to go
ahead and get in the habit of managing this verison number as we make
additions. Create a doc where we can keep a list of what changed at
each version bump.

Signed-off-by: Russell Bryant <rbryant@redhat.com>