Add test case for LLMBlock model_prompt parameter
Add a test case that ensures that passing `model_prompt` as an empty
string results in `self.model_prompt` as empty. If this parameter is
not specified (is None), `self.model_prompt` should be a non-empty
string as a result of our default behavior of picking a prompt based
on the model family.

Signed-off-by: Russell Bryant <rbryant@redhat.com>