Set gen_kwargs['n'] dynamically in the simple pipelines
We need a way to allow `--num-instructions`, or in the future
`--sdg-scale-factor`, to influence how many instructions we generate
using the simple pipelines. The way to do this seems to be to set `n`
to this value. Since this is a runtime parameter, and we only want to
set it for `n` in certain cases, add a new value for gen_kwargs['n']
called `scaled` which is a hint to use the runtime parameter here.

Closes #130

Signed-off-by: Russell Bryant <rbryant@redhat.com>