Allow mmlu_bench to return empty results
We can't control the model outputs and their ability
to produce the required format for the mmlu_bench
mcq's varys and is inconsistent.

Move the check for an Empty dataset so that we can
and use it to allow the mmlu pipeline to return an empty
dataset, while we ensure we check other pipelines for
entries.

Fixes #213

Signed-off-by: Derek Higgins <derekh@redhat.com>