checkpointing: fix "missing data" check with removed columns
Fix the logic error that means if a pipeline removes a column
that was present in the original dataset, then checkpointing
causes the column not be present before the pipeline starts.

Add a test case to the unit test to cover this, but note I
had to add an additional column to ensure there is at least
one column in common between the original dataset and the
checkpoint dataset.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>