Remove calls to logging.basicConfig on import
Fixes #196

instructlab.sdg calls logging.basicConfig on import.

Libraries should not configure logging themselves. It's the job of the
application or main script to configure logging and set up log
handlers according to their needs. logging.basicConfig is a one-shot
API. The second and any following call are a no-op, unless it is
called with force=True.

Remove setup_logging and just use logging.getLogger in the code.

Co-authored-by: Christian Heimes <cheimes@redhat.com>
Signed-off-by: Mark McLoughlin <markmc@redhat.com>