test_generate_data: adopt a much simpler no-op _generate() method
All _generate() needs to do is to produce some output based on each
input sample. We don't need to mimic with an actual LLM would produce,
anything will do in order to validate the output files.

This simplification helps avoid a buggy test lurking behind the
complexity - in this case, the test data files only contained a single
line and as a result we were only outputting a single sample.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>