Don't write empty checkpoint datasets
If we end up with empty generated datasets (perhaps because a small
teacher model was used and it was unable to generate any samples in
our expected format), we don't want to write those empty datasets out
as checkpoint files since there is no data in them to checkpoint. This
prevents us from having to write special logic on the loading side to
deal with loading empty checkpoint files.

Fixes #238

Signed-off-by: Ben Browning <bbrownin@redhat.com>