Handle empty dataset from output of sdg leaf node without raising error
Previously, an EmptyDatasetError was raised when the dataset was empty after running the sdg pipeline of a leaf node. This change logs a warning and continues processing instead, allowing the function to handle empty datasets more gracefully and process other leaf nodes in the taxonomy. Fixes https://github.com/instructlab/sdg/issues/240

Signed-off-by: Tyler Lisowski <lisowski@us.ibm.com>