Expand chunking testing, including new functional tests
This adds pytest-based functional tests to our repo with a basic PDF
chunking test.

While writing that test I discovered a minor bug in DocumentChunker,
resulting in an additional unit test and a minor change to handle the
case where we're given no documents and were previously not
instantiating any chunker.

Signed-off-by: Ben Browning <bbrownin@redhat.com>