Defer torch loading with docling until necessary
This borrows and adapts the `leanimports.py` script and test from the
InstructLab CLI repository to ensure within SDG we're not prematurely
loading the entirety of Torch into memory.

The CLI repo noticed we were doing this, and since this PR would
actually have exacerbated this by attempting to load the tesseract and
easyocr modules even earlier, this felt like the right time to address
this. The overall imports are all the same, but now we only import
specific docling pieces as needed when we're actually going to run
chunking vs triggering the whole PyTorch import chain as soon as
someone imports SDG.

Signed-off-by: Ben Browning <bbrownin@redhat.com>
(cherry picked from commit 791fc7f8b28e195fbdf77216994441b5bec96fa7)