Switch to Jinja2 templates for prompt templates
This stubs in support for Jinja templates in the LLMBlock prompt
templates, opening us up to more expressive prompts and handling things
like loops that take a variable number of input elements when
rendering templates.

NOTE: This is a backwards-incompatible change in prompt templates. Any
users that had custom pipelines specified will need to update their
template variables to look like `{{variable}}` instead of `{variable}`
as a result of this change.

Co-authored-by: shivchander <shivchander.s30@gmail.com>
Co-authored-by: abhi1092 <abhi1092@gmail.com>
Signed-off-by: Ben Browning <bbrownin@redhat.com>