Use PromptRegistry for all chat templates
This removes the mapping of model families in SDG itself between
granite, mixtral, mistral, merlinite, etc. Instead, it uses the
PromptRegistry to lookup chat templates based on the model family
given. And, if no model family is given, it still falls back to doing
a best-guess based on the file path of the selected teacher model.

A simple test was added to demonstrate how to register and use custom
chat templates for generating prompts via the PromptRegistry.

Signed-off-by: Ben Browning <bbrownin@redhat.com>