Mix all the user's generated skills instead of 30 per leaf node
Users have provided feedback that they expect all of their generated
skills to get mixed in the final output dataset by default instead of
truncating this to only 30. The choice of 30 was not a great default
for every use case anyway, and mostly tailored towards very large
taxonomies with large numbers of skills. Most of our users are using
fewer numbers of skills, and expect a larger amount of generated skill
data to make it into the results to increase the overall number of
skill samples in the mixed output that matches their custom skills.

Signed-off-by: Ben Browning <bbrownin@redhat.com>