Simplify Jinja Template undefined behavior
Only deal with setting `undefined=StrictUndefined` in a single place
for our Jinja Templates by using a Jinja Environment across all
Template instances. This is Jinja's preferred way to deal with
Templates anyway, as opposed to to automatic implicit Environment that
gets created the first time you construct a bare Template.

This also simplifies the validation logic to not deal with `ChainMap`
or any of the `__missing__` Dict stuff, and instead let Jinja deal
with that entirely for us.

Signed-off-by: Ben Browning <bbrownin@redhat.com>