Pin `docling-parse>=2.0.0,<3.0.0` + pin `pydantic==2.9.2`
`docling-parse` v3.0.0 contains breaking changes to the syntax, which is currently breaking our builds.

Also, `mypy` < v1.14  pulls in the latest version of `pydantic` by default, so this commimt hardcodes the correct span of `pydantic` versions that are compatible with `mypy` < v1.14

Finally, we want to pin `docling[tesserocr]>=2.4.2,<=2.8.3` due to breaking changes in v2.9.0 on 9 Dec 2024.

Signed-off-by: Courtney Pacheco <6019922+courtneypacheco@users.noreply.github.com>