Stub in a run_pipeline CLI and add example usage
Part of the typical research flow with SDG is to run a single
pipeline, with an input dataset and an output dataset. We need a
simple way to enable this within the SDG repo, ideally without
requiring code to be written just for the cases of running a
Pipeline. This stubs in a simple run_pipeline CLI for just that
purpose.

This also adds an example of how to use the pipeline by documenting
the IterBlock where we show how to use that block with the new
run_pipeline CLI.

Signed-off-by: Ben Browning <bbrownin@redhat.com>