Prototype a Kubeflow Pipelines version of SDG
This is a basic working prototype of our existing SDG workflow as
Kubeflow Pipelines, with an end-to-end Pipeline as well as a number of
Components for that it orchestrates for individual pieces of the
workflow. This does not yet split our SDG concept of Pipeline into a
Kubeflow Pipeline, but instead starts with just splitting up our basic
pre-processing vs generate vs post-processing.

The new `tests/functional/test_kfp_api.py` gives an example of running
both this end-to-end pipeline as well as creating a custom Kubeflow
Pipeline from our Components. The tests execute in Kubeflow Pipeline's
local mode, which means as another Python subprocess.

Signed-off-by: Ben Browning <bbrownin@redhat.com>