Use constraints-dev.txt in ilab e2e tests
This is a port of https://github.com/instructlab/instructlab/pull/3320
over to the SDG repository. While doing so, I noticed we also were not
using "-DGGML_CUDA=ON" so updated that as well, since it's the same
pip install line in the file.

Signed-off-by: Ben Browning <bbrownin@redhat.com>
(cherry picked from commit 225612cdf822c58c4e68e1d17548100015618a07)